moved ntoh, decrypt, and Decrypter into the same file
