--- src/decrypter/ntoh.js
@@ -1,12 +1,0 @@
-/**
- * Convert network-order (big-endian) bytes into their little-endian
- * representation.
- */
-const ntoh = function(word) {
-  return (word << 24) |
-    ((word & 0xff00) << 8) |
-    ((word & 0xff0000) >> 8) |
-    (word >>> 24);
-};
-
-export default ntoh;
